+++ main.js
@@ -134,8 +134,8 @@ client.on('presenceUpdate', async (oldStatus, newStatus) => {
                 }                
             }
             catch(error) {
-                console.log(`Couldn't find notification channel: ${error}`);
-                log('error', logChannel, `Couldn't find notification channel: ${error}`);
+                console.log(`Error creating streaming embed message: ${error}`);
+                log('error', logChannel, `Error creating streaming embed message: ${error}`);
             }
             // cleanup message when user stops streaming?
         }