+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] - 2021-12-
 ### Added
 - Added logging to Discord channel
+- Added checking for invalid twitch token in settings
 
 ## [0.3.0] - 2021-11-28
 ### Added